fix(snap): remove connect hooks
If the network-control and network-observe interfaces are connected before the
device service starts (for instance when they are connected by a gadget snap), then
the hooks would scan for the subnets and update the DiscoverSubnets value in Consul.

Following that, the device service would start up and the Device SDK, seeing that
the service was already present in Consul, would not add all the other
settings from the configuration.toml file. This will mean that the only setting in
Consul for the device service is the DiscoverSubnets value - and the service will not
run as expected, or be able to perform a successful device discovery.

This commit therefore removes the connect hooks.

Doing the device discovery needs to be done manually. This can be done by
following the install steps in this order:

sudo snap install edgex-device-rfid-llrp
sudo snap connect edgex-device-rfid-llrp:network-control :network-control
sudo snap connect edgex-device-rfid-llrp:network-manage :network-manage
edgex-device-rfid-llrp.discover-readers

Signed-off-by: Siggi Skulason <siggi.skulason@canonical.com>